Bunch of changes. Now has commands instead of just pasting a url.

+++ src/main.rs
@@ -1,32 +1,36 @@
+#![feature(fs_try_exists)]
+
 #[cfg(all(windows, not(debug_assertions)))]
 mod win32;
+
+mod command_handler;
+mod marco_utils;
 
 #[macro_use]
 mod printutil;
 
-use dr_downloader::{converter::Converter, downloader::Downloader, error::Result, saver::Saver};
-use std::io::{stdin, Stdin};
+use crate::command_handler::AsyncCommandHandler;
+use command_handler::Passthrough;
+use dr_downloader::{
+    cacher::{get_token, set_token},
+    converter::Converter,
+    downloader::Downloader,
+    error::Result,
+    saver::Saver,
+};
+use marco_utils::do_while;
+use std::{
+    io::{stdin, Stdin},
+    sync::Arc,
+};
+use tokio::sync::Mutex;
 
 const FFMPEG: &[u8] = include_bytes!("../ffmpeg-win32.exe");
 
-macro_rules! do_while {
-    (($cond:expr)$body:block) => {
-        loop {
-            let res = $body;
-            if !$cond {
-                break res;
-            }
-        }
-    };
-}
-
-fn clear_console() {
-    fprint!("\x1B[2J\x1B[1;1H");
-}
-
-fn log_error(err: impl AsRef<dyn std::error::Error>) {
-    fprintln!("\x1B[91mError!\x1B[0m {}", err.as_ref());
-    let err_msg = format!("{:?}", err.as_ref());
+fn log_error(err: impl AsRef<str>) {
+    let err = err.as_ref();
+    fprintln!("\x1B[91mError!\x1B[0m {}", err);
+    let err_msg = format!("{:?}", err);
     let mut file = if let Ok(f) =
         std::fs::File::open("error.txt").or_else(|_| std::fs::File::create("error.txt"))
     {
@@ -39,6 +43,9 @@ fn log_error(err: impl AsRef<dyn std::error::Error>) {
 }
 
 fn create_ffmpeg() -> Result<String> {
+    if std::fs::try_exists("ffmpeg.exe")? {
+        return Ok("ffmpeg.exe".to_owned());
+    };
     let dir = std::env::temp_dir().join("ffmpeg.exe");
     let dir_str = dir.to_string_lossy().into_owned();
     std::fs::write(&dir_str, FFMPEG)?;
@@ -89,6 +96,55 @@ async fn setup_input() -> Result<(bool, Stdin, String)> {
     Ok((input_mode, inp, input_buffer))
 }
 
+fn print_header() {
+    fprint!("\x1B[1m\x1B[47m\x1B[31m DR \x1B[49m\x1B[39m Downloader CLI\x1B[0m\n\n");
+}
+
+fn clear() {
+    fprint!("\x1B[2J\x1B[1;1H");
+    print_header();
+}
+
+async fn download(args: Vec<String>, passthrough: Passthrough) -> command_handler::Result<()> {
+    let saver = match passthrough {
+        Some(x) => Arc::downcast::<Mutex<Saver>>(x)
+            .map_err(|_| "Invalid passthrough (internal error)".to_owned())?,
+        None => return Err("Invalid passthrough (internal error)".into()),
+    };
+
+    let url = args.into_iter().collect::<String>();
+
+    let saver = saver.lock().await;
+    saver.save(url, "./").await.map_err(|e| e.to_string())?;
+
+    fprintln!("\x1B[92mDone!\x1B[0m");
+    Ok(())
+}
+
+async fn token(args: Vec<String>, _passthrough: Passthrough) -> command_handler::Result<()> {
+    let mut args_iter = args.iter();
+    let subcommand = match args_iter.next() {
+        Some(x) => x,
+        None => return Err("Expected sub command!".into()),
+    };
+    match subcommand.as_str() {
+        "set" => {
+            let value = match args_iter.next() {
+                Some(x) => x,
+                None => return Err("Expected new value!".into()),
+            };
+            set_token(value)?;
+            fprintln!("Token successfully updated!");
+        }
+        "get" => {
+            let value = get_token()?;
+            fprintln!("{}", value);
+        }
+        _ => return Err("Unknown subcommand.".into()),
+    }
+    Ok(())
+}
+
 #[tokio::main]
 async fn main() -> Result<()> {
     #[cfg(all(windows, not(debug_assertions)))]
@@ -97,26 +153,30 @@ async fn main() -> Result<()> {
     let saver = setup_saver().await?;
     let (input_mode, inp, mut input_buffer) = setup_input().await?;
 
+    let mut cmds = AsyncCommandHandler::new();
+    cmds.register("clear", |_, _| {
+        Box::pin(async {
+            clear();
+            Ok(())
+        })
+    });
+    cmds.register("exit", |_, _| panic!("Exit requested."));
+    cmds.register("download", |x, y| Box::pin(download(x.to_owned(), y)));
+    cmds.register("token", |x, y| Box::pin(token(x, y)));
+
+    let shared_saver = Arc::new(Mutex::new(saver));
+
+    print_header();
     do_while!((input_mode) {
+        input_buffer.clear();
         if input_mode {
-            clear_console();
-            fprint!("\x1B[1m\x1B[47m\x1B[31m DR \x1B[49m\x1B[39m Downloader CLI\x1B[0m\n\n");
-            fprint!("\x1B[1mEnter url:\x1B[0m ");
             inp.read_line(&mut input_buffer)?;
         }
 
-        let result = saver.save(&input_buffer, "./").await;
-
+        let result = cmds.handle(&input_buffer, Some(shared_saver.clone())).await;
         if let Err(val) = result {
-            log_error(val);
-            const CLEAR_TIME: u16 = 5000;
-            fprint!("Clearing in {}s", CLEAR_TIME / 1000);
-            std::thread::sleep(std::time::Duration::from_millis(CLEAR_TIME as u64));
-            continue;
+            log_error(val.to_string());
         }
-
-        fprintln!("\x1B[92mDone!\x1B[0m");
-        std::thread::sleep(std::time::Duration::from_millis(2000));
     });
     Ok(())
 }
